feat(end-screen): adjust navigators to the new architecture
