+++ App.js
@@ -1,21 +1,13 @@
 import { StatusBar } from 'expo-status-bar';
-import { createDrawerNavigator } from '@react-navigation/drawer';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import { NavigationContainer } from '@react-navigation/native';
 import { Provider } from 'react-redux';
-import Settings from './settings.json';
 // providers and store
 import { store } from './src/store/redux/store';
-// components
-import CustomDrawer from './src/components/custom-drawer';
-import IconButton from './src/components/icon-button';
 // screens
-import PollNavigator from './src/screens/poll-screen/navigator';
-import ClassroomMainScreen from './src/screens/classroom-main-screen';
-import UserParticipantsScreen from './src/screens/user-participants-screen';
+import DrawerNavigator from './src/components/custom-drawer/drawer-navigator';
 import FullscreenWrapper from './src/components/fullscreen-wrapper';
-import TestComponentsScreen from './src/screens/test-components-screen';
-import UserNotesScreen from './src/screens/user-notes-screen';
-import WhiteboardScreen from './src/screens/whiteboard-screen';
+import EndSessionScreen from './src/screens/end-session-screen';
 import { injectStore as injectStoreVM } from './src/services/webrtc/video-manager';
 import { injectStore as injectStoreSM } from './src/services/webrtc/screenshare-manager';
 import { injectStore as injectStoreAM } from './src/services/webrtc/audio-manager';
@@ -31,127 +23,24 @@ const injectStore = () => {
 
 const App = ({ onLeaveSession, jUrl }) => {
   injectStore();
-  const Drawer = createDrawerNavigator();
+  const Stack = createNativeStackNavigator();
 
   return (
     <>
       <Provider store={store}>
         <NavigationContainer independent>
-          {!Settings.dev && <TestComponentsScreen jUrl={jUrl} />}
-          <Drawer.Navigator
-            independent
-            drawerContent={(props) => <CustomDrawer {...props} onLeaveSession={onLeaveSession} />}
+          <Stack.Navigator
             screenOptions={{
-              contentOptions: {
-                style: {
-                  backgroundColor: 'black',
-                  flex: 1,
-                },
-              },
-
-              sceneContainerStyle: { backgroundColor: '#06172A' },
-              drawerActiveBackgroundColor: '#003399',
-              drawerActiveTintColor: 'white',
-              headerStyle: { backgroundColor: '#003399' },
-              headerTintColor: 'white',
-              drawerBackgroundColor: '#003399',
-              headerTitleAlign: 'center',
+              headerShown: false,
+              contentStyle: { backgroundColor: '#06172A' }
             }}
           >
-            <Drawer.Screen
-              name="Main"
-              component={ClassroomMainScreen}
-              options={{
-                title: 'Sala de aula',
-                drawerIcon: (config) => (
-                  <IconButton
-                    icon="home"
-                    size={18}
-                    iconColor={config.color}
-                  />
-                ),
-              }}
-            />
-
-            <Drawer.Screen
-              name="SharedNoteScreen"
-              component={UserNotesScreen}
-              options={{
-                title: 'Nota compartilhada',
-                drawerIcon: (config) => (
-                  <IconButton
-                    icon="file-document"
-                    size={18}
-                    iconColor={config.color}
-                  />
-                ),
-              }}
-            />
-
-            <Drawer.Screen
-              name="PollScreen"
-              component={PollNavigator}
-              options={{
-                title: 'Enquete',
-                drawerIcon: (config) => (
-                  <IconButton
-                    icon="chart-box"
-                    size={18}
-                    iconColor={config.color}
-                  />
-                ),
-              }}
-            />
-
-            <Drawer.Screen
-              name="UserParticipantsScreen"
-              component={UserParticipantsScreen}
-              options={{
-                title: 'Lista de participantes',
-                drawerIcon: (config) => (
-                  <IconButton
-                    icon="account-multiple"
-                    size={18}
-                    iconColor={config.color}
-                  />
-                ),
-              }}
-            />
-
-            <Drawer.Screen
-              name="WhiteboardScreen"
-              component={WhiteboardScreen}
-              options={{
-                title: 'Quadro Branco',
-                drawerIcon: (config) => (
-                  <IconButton
-                    icon="brush"
-                    size={18}
-                    iconColor={config.color}
-                  />
-                ),
-              }}
-            />
-
-            {Settings.dev && (
-            <Drawer.Screen
-              name="TestComponent"
-              options={{
-                title: 'Test Component',
-                drawerIcon: (config) => (
-                  <IconButton
-                    icon="brush"
-                    size={18}
-                    iconColor={config.color}
-                  />
-                ),
-              }}
-            >
-              {(props) => <TestComponentsScreen {...props} jUrl={jUrl} />}
-            </Drawer.Screen>
-            )}
-
-          </Drawer.Navigator>
+            {/*<Stack.Screen name="GuestScreen" component={GuestScreen} />*/}
+            <Stack.Screen name="DrawerNavigator">
+              {() => <DrawerNavigator jUrl={jUrl} onLeaveSession={onLeaveSession} />}
+            </Stack.Screen>
+            <Stack.Screen name="EndSessionScreen" component={EndSessionScreen} />
+          </Stack.Navigator>
         </NavigationContainer>
         <FullscreenWrapper />
       </Provider>
